% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_get.R
\name{.download_dependency_cmd}
\alias{.download_dependency_cmd}
\title{Build the command to have maven download a dependency (package-private)}
\usage{
.download_dependency_cmd(dep, group, version, mvn, transitive = TRUE)
}
\arguments{
\item{dep}{Character. The dependency name. This could be the entire name (e.g.,
group:dependency:version) in which case the other inputs may be omitted. Otherwise
it can be just the dependency name (i.e., the part in the middle of the above example).}

\item{group}{Character. The maven artifact group name.}

\item{version}{Character. The desired version of the dependency.}

\item{mvn}{Character. The path the the maven installation.}

\item{transitive}{Logical. If \code{TRUE}, download transitively, retrieving the specified artifact and all of its dependencies.}
}
\description{
This function will build a system command using the maven dependency get funtionality to download a (java)
dependency.
}
\keyword{internal}
