% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_get.R
\name{download_dependency}
\alias{download_dependency}
\title{Have maven download a dependency}
\usage{
download_dependency(dep, group, version, mvn = find_mvn(), java_home)
}
\arguments{
\item{dep}{Character. The dependency name. This could be the entire name (e.g.,
group:dependency:version) in which case the other inputs may be omitted. Otherwise
it can be just the dependency name (i.e., the part in the middle of the above example).}

\item{group}{Character. The maven artifact group name.}

\item{version}{Character. The desired version of the dependency.}

\item{mvn}{Character. The path the the maven installation.}

\item{m2}{Character. Path to local maven repository. If no input is provided a standard
location will be guessed at (platform dependent).}
}
\description{
This function will use the maven dependency get funtionality to download a (java)
dependency.
}

