% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_get.R
\name{find_artifacts}
\alias{find_artifacts}
\title{Have maven find a dependency}
\usage{
find_artifacts(artifacts, verbose = FALSE)
}
\description{
This function will find a maven dependency in your local maven repository. Note that
if the dependency has not yet been downloaded (e.g., via \code{\link{download_dependency}})
then the return value will be \code{NULL}.
}
\seealso{
\link{get_dependencies}
}
