% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maven.R
\name{find_mvn}
\alias{find_mvn}
\alias{use_mvn}
\title{Find maven on the system}
\usage{
find_mvn()

use_mvn(mvn)
}
\arguments{
\item{mvn}{Character. A path to a maven executable}
}
\description{
Determine where maven is installed.
}
\details{
There are a few options for setting which maven install to use. If maven is installed in
a standard location then \code{find_mvn} will probably work with no help. Users may also
choose a maven version by setting a \code{maven.path} option or by calling the
\code{use_mvn} function. Once called, the package will remember the setting until detached.
}
