% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_classpath.R
\name{get_classpath_from_pom}
\alias{get_classpath_from_pom}
\title{Get a java classpath from POM}
\usage{
get_classpath_from_pom(args, cp_file, verbose = FALSE)
}
\arguments{
\item{path}{Character. Path to the directory containing a pom.xml to derive paths for}
}
\description{
This function requires a simple pom.xml to function, so its usability is limited. However it can
be very handy for packages with java dependencies where there is not a fat jar to depend upon.
If you have a list of dependencies but no actual pom.xml, see \code{\link{get_classpath}}
}
