% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{set_mvn_repos}
\alias{set_mvn_repos}
\alias{add_mvn_repos}
\title{Set/Append Maven Artifact Repositories}
\usage{
set_mvn_repos(repos)

add_mvn_repos(repos, prepend = FALSE)
}
\arguments{
\item{repos}{Character vector. Should be maven repo URLs}

\item{prepend}{Logical. If \code{TRUE} the new paths will come before the old.}
}
\description{
\code{set_mvn_repos} will overwrite any existing repos and
\code{add_mvn_repos} will append/prepend to existing.
}
\details{
These repositories are only used by the \code{dependency:get} maven plugin, so they
will be used when calling \code{\link{get_dependencies}} but not when calling
\code{\link{build_classpath}} (which uses the \code{dependency:build-classpath} maven
plugin under the hood). To set repositories for building classpaths it is necessary
to update the maven settings.xml file. This is typically found in \code{~/.m2/settings.xml}.
}
