% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maven.R
\name{test_mvn}
\alias{test_mvn}
\title{Test Maven}
\usage{
test_mvn(mvn, verbose = FALSE)
}
\value{
Logical. If \code{TRUE} maven is working fine. Otherwise it isn't.
}
\description{
Test to ensure that maven is working on your system by trying to run a command
to get the maven version you're running.
}
