% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_pom.R
\name{write_pom}
\alias{write_pom}
\title{Write a basic pom.xml}
\usage{
write_pom(dep, group, version, mvn = find_mvn(), java_home,
  transitive = TRUE, path = tempdir(), artifactid = "tempproj",
  groupid = "tempgroup", artifactversion = "0.0.1-SNAPSHOT")
}
\arguments{
\item{path}{Character. The path where the pom should be written.}

\item{artifactid}{Character. The \code{artifactId} entry in the resulting pom.xml. Unless you intend
to save this pom for future use this entry does not really matter and the default is fine.}

\item{groupid}{Character. The \code{groupId} entry in the resulting pom.xml. See \code{artifactid}
for use case.}

\item{artifactversion}{Character. The \code{version} entry in the resulting pom.xml. See \code{artifactid}
for use case.}
}
\value{
the file path of the newly written pom.xml
}
\description{
The purpose of this function is to write a pom.xml to support classpath management when using rJava on
dependencies with dependencies of their own. For the most part this is probably not a function that users
should be calling directly.
}

