% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_get.R
\name{download_dependency}
\alias{download_dependency}
\title{Have maven download a dependency}
\usage{
download_dependency(dep, group, version, mvn = find_mvn(), java_home,
  transitive = TRUE, quiet = FALSE, dep_plugin_version = "3.1.1")
}
\arguments{
\item{dep}{Character. The dependency name. This could be the entire name (e.g.,
group:dependency:version) in which case the other inputs may be omitted. Otherwise
it can be just the dependency name (i.e., the part in the middle of the above example).}

\item{group}{Character. The maven artifact group name.}

\item{version}{Character. The desired version of the dependency.}

\item{mvn}{Character. The path the the maven installation.}

\item{java_home}{Character. Path to java. If not provided the standard install paths
(platform dependent) will be checked.}

\item{transitive}{Logical. If \code{TRUE}, download transitively, retrieving the specified artifact and all of its dependencies.}

\item{quiet}{Logical. If \code{FALSE} status messages and some logs/warnings will be printed to the console.}

\item{dep_plugin_version}{Character. The version of the maven dependency plugin. See https://github.com/apache/maven-dependency-plugin/releases
for release versions.}
}
\description{
This function will use the maven dependency get funtionality to download a (java)
dependency.
}
