% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_get.R
\name{find_artifacts}
\alias{find_artifacts}
\title{Have maven find a dependency}
\usage{
find_artifacts(artifacts, resolve = FALSE, verbose = FALSE)
}
\arguments{
\item{artifacts}{Character vector. Should be of the form 'groupId:artifactId:version[:packaging[:classifier]]'
where the packaging and classifier inputs are optional.}

\item{resolve}{Logical. If \code{TRUE} use \code{get_dependencies} to resolve any requested
artifacts that are not found in the local maven repository.}

\item{verbose}{Logical. Dump maven output to console.}
}
\description{
This function will find a maven dependency in your local maven repository. Note that
if the dependency has not yet been downloaded (e.g., via \code{\link{download_dependency}})
then the return value will be \code{NULL}.
}
\seealso{
\link{get_dependencies}
}
