% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maven.R
\name{find_local_mvn_repo}
\alias{find_local_mvn_repo}
\title{Find your local maven repository}
\usage{
find_local_mvn_repo(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. Dump maven output}
}
\description{
This is primarily intended for internal use. It may however be useful for some to have
a programatic way of identifying the local maven reposiory.
}
