% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{find_mvn}
\alias{find_mvn}
\title{Find maven on the system}
\usage{
find_mvn()
}
\description{
The package uses maven to resolve dependencies. It does not come packaged with
maven. As such it is required that the user independently install maven. This
function will try to find the maven executable assuming that maven is already
installed.
}
