% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_cp.R
\name{get_classpath_from_pom}
\alias{get_classpath_from_pom}
\title{Get a java classpath from POM}
\usage{
get_classpath_from_pom(path, mvn = find_mvn(), java_home,
  quiet = FALSE)
}
\arguments{
\item{path}{Character. Path to the directory containing a pom.xml to derive paths for}

\item{mvn}{Character. The path the the maven installation.}

\item{java_home}{Character. Path to java. If not provided the standard install paths
(platform dependent) will be checked.}

\item{quiet}{Logical. If \code{FALSE} status messages and some logs/warnings will be printed to the console.}
}
\description{
This function requires a simple pom.xml to function, so its usability is limited. However it can
be very handy for packages with java dependencies where there is not a fat jar to depend upon.
If you have a list of dependencies but no actual pom.xml, see \code{\link{get_dependency_path}}
}
