% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{set_java_home}
\alias{clear_java_home}
\alias{set_java_home}
\title{Set/clear JAVA_HOME}
\usage{
set_java_home(java_home)

clear_java_home()
}
\arguments{
\item{java_home}{Character. Indicates location of java on the system. If
not provided then the function will attempt to find java using standard
install paths.}
}
\value{
Logical. \code{TRUE} if JAVA_HOME was set and \code{FALSE} otherwise.
}
\description{
Maven needs to know where java is to run. Setting the JAVA_HOME
environment variable can however cause issues with \code{rJava`}. If
issues do arise use \code{clear_java_home} after your dependencies are
resolved.
}

