% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java.R
\name{set_java_home}
\alias{clear_java_home}
\alias{find_java}
\alias{is_java_available}
\alias{set_java_home}
\title{Set/clear JAVA_HOME}
\usage{
set_java_home(java_home = find_java(), quiet = FALSE)

find_java()

clear_java_home()

is_java_available()
}
\arguments{
\item{java_home}{Character. Indicates location of java on the system. If
not provided then the function will attempt to find java using standard
install paths.}

\item{quiet}{Logical. If \code{FALSE} the new JAVA_HOME setting will be printed.}
}
\description{
Maven needs to know where java is to run. Setting the JAVA_HOME
environment variable can however cause issues with \code{rJava}. If
issues do arise use \code{clear_java_home} after your dependencies are
resolved.
}
\section{java availability}{


This function runs \code{mvn -version} to determine if java is on the system path used
by maven when executed from R. This test is a little different from, for example, 
executing \code{system("java -version")}. There are cases where java may be available
for system commands but not specifically set on a system path and hence unavailable for
use by maven.
}

