% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java.R
\name{set_java_home}
\alias{set_java_home}
\alias{find_java}
\alias{clear_java_home}
\title{Java for Maven}
\usage{
set_java_home(java_home = find_java(), quiet = FALSE)

find_java()

clear_java_home(quiet = FALSE)
}
\arguments{
\item{java_home}{Character. Indicates location of java on the system. If
not provided then the function will attempt to find java using standard
install paths.}

\item{quiet}{Logical. If \code{FALSE} the new JAVA_HOME setting will be printed.}
}
\description{
Maven needs java to run. In this case it means that java needs to be available
for maven system commands.
}
\section{set/clear JAVA_HOME}{


Maven needs to know where java is to run. Setting the JAVA_HOME
environment variable can however cause issues with \code{rJava}. If
issues do arise use \code{clear_java_home} after your dependencies are
resolved.
}

\section{find java}{


Find a JRE/JDK on the system. This is a convenience function to find java
on the system. It will use \code{getOption(".java.home")} if that is found.
Otherwise it will look in standard install locations depending on the
system platform.
}

